done changes for division changes from int to double

+++ Operation/src/com/ivl/MathOperation.java
@@ -9,7 +9,7 @@ public class MathOperation {
 		System.out.println("substraction :"+b);
 		int m= 190*90;
 		System.out.println("multiplication :"+m);
-		int d= 190/90;
+		double d= 190/90;
 		System.out.println("division :"+d);
